actuarial-fastapi/requirements.txt
# FastAPI service requirements
fastapi>=0.95
uvicorn[standard]>=0.21
pydantic>=1.10
httpx>=0.24
python-json-logger>=2.0
python-multipart>=0.0.6
PyJWT>=2.8
python-dotenv>=1.0
orjson>=3.8.6          # optional: faster JSON if you use it in FastAPI
requests>=2.31.0       # used by sync proxy or tests
pytest>=7.0
pytest-asyncio>=0.21


